Rename jars
This renames the output jars:
Arctic.jar -> arctic-<VERSION>.jar
ArcticShared.jar -> arctic-api-<VERSION>.jar
