+++ README.md
@@ -37,11 +37,11 @@ As **Arctic** relies on pixel comparison to validate the different results, chan
 
 Because **Arctic** captures all keyboard and mouse events, during the recording we need a way to tell **Arctic** what we want it to do without interfering with the test. This can be done by pressing specific key combination. When several modifier keys are pressed (such as **ctrl** or **alt**) **Arctic** will interpret the next key presses as instructions. Among others, these instructions can be to start or stop the recording or to capture a screenshot.
 
-The keys are configured in the **recorder.properties** file with properties that start with **arctic.recorder.control.jnh**. The correspondence between keycodes and keys can be checked by running ```java -jar Arctic.jar -k```. This command will print the keycode of the different keys as they are pressed. By default, **ctrl+alt+z** is used to start/stop the recording and **ctrl+alt+x** is used to capture a screenshot.
+The keys are configured in the **recorder.properties** file with properties that start with **arctic.recorder.control.jnh**. The correspondence between keycodes and keys can be checked by running ```java -jar arctic-<VERSION>.jar -k```. This command will print the keycode of the different keys as they are pressed. By default, **ctrl+alt+z** is used to start/stop the recording and **ctrl+alt+x** is used to capture a screenshot.
 
 ### Record a test
 
-Before running a test, start **Arctic** in recording mode by executing ```java -jar Arctic.jar -r```.
+Before running a test, start **Arctic** in recording mode by executing ```java -jar arctic-<VERSION>.jar -r```.
 
 ![Workbench](images/workbench.png)
 
@@ -49,7 +49,7 @@ Once you see the *Workbench* and *Shade windows*, you can start running a test c
 
 Additionally, we need to tell **Arctic** which test is on the screen. Arctic identifies tests with two strings, the **testName** and the **testCase**. We can do this by writing test run **<testName> <testCase>** in the arctic command line.
 
-**Arctic** can also receive this information via RMI. This requires setting the configuration key **arctic.common.cmd.rmi.enabled** from **false** to either **local_only** or **true**. Once this is done, we can send this information through a second arctic instance, running ```java -jar Arctic.jar -c test run <testName> <testCase>``` or by implementing a custom application that uses the RMI interface present in **ArcticShared.jar**
+**Arctic** can also receive this information via RMI. This requires setting the configuration key **arctic.common.cmd.rmi.enabled** from **false** to either **local_only** or **true**. Once this is done, we can send this information through a second arctic instance, running ```java -jar arctic-<VERSION>.jar -c test run <testName> <testCase>``` or by implementing a custom application that uses the RMI interface present in **ArcticShared.jar**
 
 ![simplebuttontest](images/simplebuttontest.png)
 
@@ -73,18 +73,18 @@ Screenshots and recordings will be saved in the **arctic_tests** folder. This ca
 
 ### Replay a test
 
-Once the recording is finished, **Arctic** is ready to replay the test. Launch **Arctic** in player mode with ```java -jar Arctic.jar -p```.
+Once the recording is finished, **Arctic** is ready to replay the test. Launch **Arctic** in player mode with ```java -jar arctic-<VERSION>.jar -p```.
 
-Once **Arctic** has loaded, start the relevant test and then instruct **Arctic** to run it with ```java -jar Arctic.jar -c test run <testName> <testCase>```. **Arctic** will reproduce the keyboard and mouse events and capture screenshots at the same time as in the recording.
+Once **Arctic** has loaded, start the relevant test and then instruct **Arctic** to run it with ```java -jar arctic-<VERSION>.jar -c test run <testName> <testCase>```. **Arctic** will reproduce the keyboard and mouse events and capture screenshots at the same time as in the recording.
 
 <video src="https://github.com/user-attachments/assets/ae7f6bd1-5c18-4bff-8e30-ce118c6150fb" controls="controls" style="max-width: 730px;">
 </video>
 
-Once the test is finished, we can pass this information to **Arctic** with ```java -jar Arctic.jar test finish <testName> <testCase> <result>```. We can use result to tell **Arctic** whether anything has gone wrong on the test side, for example, based on the return code of the test. If result is true and all screenshots captured during the test are considered good enough, **Arctic** will mark the test as OK. If any of the screenshots were different or the result is false, **Arctic** will mark the test as failed.
+Once the test is finished, we can pass this information to **Arctic** with ```java -jar arctic-<VERSION>.jar test finish <testName> <testCase> <result>```. We can use result to tell **Arctic** whether anything has gone wrong on the test side, for example, based on the return code of the test. If result is true and all screenshots captured during the test are considered good enough, **Arctic** will mark the test as OK. If any of the screenshots were different or the result is false, **Arctic** will mark the test as failed.
 
 ### Getting Arctic results
 
-Once all the tests have run, we can export the results into one of the supported formats  **junit**, **tap**, or **jtx**. Such report files that will contain a list of the tests that **passed/failed** can be generated with ```java -jar Arctic.jar <format> save <filename>```. For example, generated jtx files for failed tests can later be used as exclude files for jtHarness.
+Once all the tests have run, we can export the results into one of the supported formats  **junit**, **tap**, or **jtx**. Such report files that will contain a list of the tests that **passed/failed** can be generated with ```java -jar arctic-<VERSION>.jar <format> save <filename>```. For example, generated jtx files for failed tests can later be used as exclude files for jtHarness.
 Valid formats are:
 
 - **xml**: junit xml report
@@ -93,7 +93,7 @@ Valid formats are:
 
 ### Review screenshot comparisons
 
-For multiple reasons, tests may not pass for the first time after recording. The results can be reviewed by using the **Arctic** command sc. For example, ```java -jar Arctic.jar -c sc all``` will start a review of all the screenshots that failed in the current session. This can help you identify why the recording failed, and also add the current screenshot as an alternative to be compared in future runs.
+For multiple reasons, tests may not pass for the first time after recording. The results can be reviewed by using the **Arctic** command sc. For example, ```java -jar arctic-<VERSION>.jar -c sc all``` will start a review of all the screenshots that failed in the current session. This can help you identify why the recording failed, and also add the current screenshot as an alternative to be compared in future runs.
 
 ### Contribution Guidelines Instructions
