+++ cmd_client/src/main/java/com/amazon/corretto/arctic/cmd/client/rmi/impl/ClientRmiInterceptor.java
@@ -19,7 +19,7 @@ package com.amazon.corretto.arctic.cmd.client.rmi.impl;
 import javax.inject.Inject;
 
 import com.amazon.corretto.arctic.common.command.impl.BaseCommand;
-import com.amazon.corretto.arctic.shared.rmi.ArcticRmiCommandClient;
+import com.amazon.corretto.arctic.api.rmi.ArcticRmiCommandClient;
 
 /**
  * Most commands are executed via RMI, contacting the ArcticPlayer or ArcticRecorder running. But some commands can be
