+++ common/src/main/java/com/amazon/corretto/arctic/common/backend/impl/JavaImageIoSaver.java
@@ -22,7 +22,7 @@ import javax.imageio.ImageIO;
 import javax.inject.Inject;
 
 import com.amazon.corretto.arctic.common.backend.ArcticImageSaver;
-import com.amazon.corretto.arctic.shared.exception.ArcticException;
+import com.amazon.corretto.arctic.api.exception.ArcticException;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
