+++ common/src/main/java/com/amazon/corretto/arctic/common/repository/impl/JsonFileTestSaveRepositoryImpl.java
@@ -40,7 +40,7 @@ import com.amazon.corretto.arctic.common.model.TestId;
 import com.amazon.corretto.arctic.common.model.event.ScreenshotCheck;
 import com.amazon.corretto.arctic.common.repository.TestSaveRepository;
 import com.amazon.corretto.arctic.common.util.Pair;
-import com.amazon.corretto.arctic.shared.exception.ArcticException;
+import com.amazon.corretto.arctic.api.exception.ArcticException;
 import com.google.gson.Gson;
 import lombok.extern.slf4j.Slf4j;
