+++ player/src/main/java/com/amazon/corretto/arctic/player/backend/impl/ArcticImageCheckPlayer.java
@@ -27,7 +27,7 @@ import com.amazon.corretto.arctic.player.backend.ArcticBackendPlayer;
 import com.amazon.corretto.arctic.player.backend.ImageComparator;
 import com.amazon.corretto.arctic.player.control.TimeController;
 import com.amazon.corretto.arctic.player.model.ArcticRunningTest;
-import com.amazon.corretto.arctic.shared.exception.ArcticException;
+import com.amazon.corretto.arctic.api.exception.ArcticException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
