+++ player/src/main/java/com/amazon/corretto/arctic/player/control/impl/AdvancedTimeController.java
@@ -26,7 +26,7 @@ import com.amazon.corretto.arctic.common.model.event.ArcticEvent;
 import com.amazon.corretto.arctic.common.tweak.ArcticTweakableComponent;
 import com.amazon.corretto.arctic.common.tweak.TweakKeys;
 import com.amazon.corretto.arctic.player.control.TimeController;
-import com.amazon.corretto.arctic.shared.exception.ArcticException;
+import com.amazon.corretto.arctic.api.exception.ArcticException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
