+++ player/src/main/resources/player.properties
@@ -118,7 +118,7 @@ arctic.player.backend.jnhMouse.events = 0x03700
 arctic.player.backend.awtKeyboard.keymap.bundled = true
 
 # Bundled resource name or path to the file with the keymap.
-# These files can be generated running `java -jar Arctic.jar -d`.
+# These files can be generated running `java -jar arctic-<VERSION>.jar -d`.
 arctic.player.backend.awtKeyboard.keymap = keymap_linux.txt
 
 # Define which image comparator to use when reproducing the tests
