+++ recorder/src/main/java/com/amazon/corretto/arctic/recorder/postprocessing/impl/ScreenCheckHashPostProcessor.java
@@ -24,7 +24,7 @@ import com.amazon.corretto.arctic.common.model.ArcticTest;
 import com.amazon.corretto.arctic.common.model.event.ScreenshotCheck;
 import com.amazon.corretto.arctic.recorder.inject.InjectionKeys;
 import com.amazon.corretto.arctic.recorder.postprocessing.ArcticRecorderPostProcessor;
-import com.amazon.corretto.arctic.shared.exception.ArcticException;
+import com.amazon.corretto.arctic.api.exception.ArcticException;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
