+++ recorder/src/main/java/com/amazon/corretto/arctic/recorder/preprocessing/impl/FirstScCheckPreProcessor.java
@@ -26,7 +26,7 @@ import com.amazon.corretto.arctic.common.model.event.ScreenshotCheck;
 import com.amazon.corretto.arctic.common.model.gui.ScreenArea;
 import com.amazon.corretto.arctic.recorder.inject.InjectionKeys;
 import com.amazon.corretto.arctic.recorder.preprocessing.ArcticRecorderPreProcessor;
-import com.amazon.corretto.arctic.shared.exception.ArcticException;
+import com.amazon.corretto.arctic.api.exception.ArcticException;
 
 public final class FirstScCheckPreProcessor implements ArcticRecorderPreProcessor {
     private static final org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(FirstScCheckPreProcessor.class);
